+++ ntex-service/src/cfg.rs
@@ -1,40 +1,27 @@
 //! Shared configuration for services
 #![allow(clippy::should_implement_trait, clippy::new_ret_no_self)]
+use std::any::{Any, TypeId};
 use std::cell::{RefCell, UnsafeCell};
-use std::sync::atomic::{AtomicUsize, Ordering};
-use std::{any::Any, any::TypeId, fmt, marker::PhantomData, ops, ptr, rc};
+use std::sync::{Arc, atomic::AtomicUsize, atomic::Ordering};
+use std::{fmt, hash::Hash, hash::Hasher, marker::PhantomData, mem, ops, ptr, rc};
 
 type Key = (usize, TypeId);
 type HashMap<K, V> = std::collections::HashMap<K, V, foldhash::fast::RandomState>;
+type HashSet<V> = std::collections::HashSet<V, foldhash::fast::RandomState>;
 
 thread_local! {
-    static DEFAULT_CFG: &'static Storage = Box::leak(Box::new(
-        Storage::new("--".to_string(), false)));
-    static MAPPING: RefCell<HashMap<Key, &'static dyn Any>> = {
+    static DEFAULT_CFG: Arc<Storage> = {
+        let mut st = Arc::new(Storage::new("--", false));
+        let ctx = CfgContext(Arc::as_ptr(&st));
+        Arc::get_mut(&mut st).unwrap().ctx = ctx;
+        st
+    };
+    static MAPPING: RefCell<HashMap<Key, Arc<dyn Any + Send + Sync>>> = {
         RefCell::new(HashMap::default())
     };
+    static REFS: RefCell<HashSet<SharedCfg>> = RefCell::new(HashSet::default());
 }
 static IDX: AtomicUsize = AtomicUsize::new(0);
-
-#[derive(Debug)]
-struct Storage {
-    id: usize,
-    tag: String,
-    building: bool,
-    data: HashMap<TypeId, Box<dyn Any + Send + Sync>>,
-}
-
-impl Storage {
-    fn new(tag: String, building: bool) -> Self {
-        let id = IDX.fetch_add(1, Ordering::SeqCst);
-        Storage {
-            id,
-            tag,
-            building,
-            data: HashMap::default(),
-        }
-    }
-}
 
 pub trait Configuration: Default + Send + Sync + fmt::Debug + 'static {
     const NAME: &'static str;
@@ -44,20 +31,45 @@ pub trait Configuration: Default + Send + Sync + fmt::Debug + 'static {
     fn set_ctx(&mut self, ctx: CfgContext);
 }
 
-#[derive(Clone, Debug)]
-pub struct CfgContext(&'static Storage);
+#[derive(Debug)]
+struct Storage {
+    id: usize,
+    tag: &'static str,
+    ctx: CfgContext,
+    building: bool,
+    data: HashMap<TypeId, Arc<dyn Any + Send + Sync>>,
+}
+
+impl Storage {
+    fn new(tag: &'static str, building: bool) -> Self {
+        let id = IDX.fetch_add(1, Ordering::SeqCst);
+        Storage {
+            id,
+            tag,
+            building,
+            data: HashMap::default(),
+            ctx: CfgContext(ptr::null()),
+        }
+    }
+}
+
+#[derive(Debug)]
+pub struct CfgContext(*const Storage);
+
+unsafe impl Send for CfgContext {}
+unsafe impl Sync for CfgContext {}
 
 impl CfgContext {
     #[inline]
     /// Unique id of the context.
     pub fn id(&self) -> usize {
-        self.0.id
+        self.get_ref().id
     }
 
     #[inline]
     /// Context tag
     pub fn tag(&self) -> &'static str {
-        self.0.tag.as_ref()
+        self.get_ref().tag
     }
 
     #[inline]
@@ -66,31 +78,38 @@ impl CfgContext {
     where
         T: Configuration,
     {
-        SharedCfg(self.0).get()
+        get(self.get_ref())
     }
 
     #[inline]
     /// Get a shared configuration.
     pub fn shared(&self) -> SharedCfg {
-        SharedCfg(self.0)
+        let inner: Arc<Storage> = unsafe { Arc::from_raw(self.0) };
+        let shared = SharedCfg(inner.clone());
+        mem::forget(inner);
+        shared
+    }
+
+    fn get_ref(&self) -> &Storage {
+        unsafe { self.0.as_ref().unwrap() }
     }
 }
 
 impl Default for CfgContext {
     #[inline]
     fn default() -> Self {
-        CfgContext(DEFAULT_CFG.with(|cfg| *cfg))
+        CfgContext(DEFAULT_CFG.with(Arc::as_ptr))
     }
 }
 
 #[derive(Debug)]
-pub struct Cfg<T: Configuration>(UnsafeCell<&'static T>, PhantomData<rc::Rc<str>>);
+pub struct Cfg<T: Configuration>(UnsafeCell<*const T>, PhantomData<rc::Rc<T>>);
 
 impl<T: Configuration> Cfg<T> {
     #[inline]
     /// Unique id of the configuration.
     pub fn id(&self) -> usize {
-        self.get_ref().ctx().0.id
+        self.get_ref().ctx().id()
     }
 
     #[inline]
@@ -106,7 +125,7 @@ impl<T: Configuration> Cfg<T> {
     }
 
     fn get_ref(&self) -> &T {
-        unsafe { &*self.0.get() }
+        unsafe { (*self.0.get()).as_ref().unwrap() }
     }
 
     /// Replaces the inner value.
@@ -122,17 +141,29 @@ impl<T: Configuration> Cfg<T> {
     }
 }
 
+impl<T: Configuration> Drop for Cfg<T> {
+    fn drop(&mut self) {
+        unsafe { drop(Arc::<T>::from_raw(*self.0.get())) }
+    }
+}
+
 impl<T: Configuration> Clone for Cfg<T> {
     #[inline]
     fn clone(&self) -> Self {
-        Cfg(UnsafeCell::new(unsafe { &*self.0.get() }), PhantomData)
+        let cloned = unsafe {
+            let inner = Arc::<T>::from_raw(*self.0.get());
+            let cloned = inner.clone();
+            mem::forget(inner);
+            Arc::into_raw(cloned).cast_mut()
+        };
+        Cfg(UnsafeCell::new(cloned), PhantomData)
     }
 }
 
 impl<'a, T: Configuration> From<&'a T> for Cfg<T> {
     #[inline]
     fn from(cfg: &'a T) -> Self {
-        SharedCfg(cfg.ctx().0).get()
+        get(cfg.ctx().get_ref())
     }
 }
 
@@ -154,26 +185,32 @@ impl<T: Configuration> Default for Cfg<T> {
 
 #[derive(Clone, Debug)]
 /// Shared configuration
-pub struct SharedCfg(&'static Storage);
+pub struct SharedCfg(Arc<Storage>);
 
 #[derive(Debug)]
 pub struct SharedCfgBuilder {
     ctx: CfgContext,
-    storage: Option<Box<Storage>>,
+    storage: Arc<Storage>,
 }
 
 impl Eq for SharedCfg {}
 
 impl PartialEq for SharedCfg {
     fn eq(&self, other: &Self) -> bool {
-        ptr::from_ref(self.0) == ptr::from_ref(other.0)
+        ptr::from_ref(self.0.as_ref()) == ptr::from_ref(other.0.as_ref())
+    }
+}
+
+impl Hash for SharedCfg {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        ptr::from_ref(self.0.as_ref()).hash(state);
     }
 }
 
 impl SharedCfg {
     /// Construct new configuration
-    pub fn new<T: AsRef<str>>(tag: T) -> SharedCfgBuilder {
-        SharedCfgBuilder::new(tag.as_ref().to_string())
+    pub fn new(tag: &'static str) -> SharedCfgBuilder {
+        SharedCfgBuilder::new(tag)
     }
 
     #[inline]
@@ -185,7 +222,7 @@ impl SharedCfg {
     #[inline]
     /// Get tag
     pub fn tag(&self) -> &'static str {
-        self.0.tag.as_ref()
+        self.0.tag
     }
 
     /// Get a reference to a previously inserted on configuration.
@@ -196,55 +233,14 @@ impl SharedCfg {
     where
         T: Configuration,
     {
-        assert!(
-            !self.0.building,
-            "{}: Cannot access shared config while building",
-            self.tag()
-        );
-
-        let tp = TypeId::of::<T>();
-        self.0
-            .data
-            .get(&tp)
-            .and_then(|boxed| boxed.downcast_ref())
-            .map_or_else(
-                || {
-                    MAPPING.with(|store| {
-                        let key = (self.0.id, tp);
-                        if let Some(boxed) = store.borrow().get(&key) {
-                            Cfg(UnsafeCell::new(boxed.downcast_ref().unwrap()), PhantomData)
-                        } else {
-                            log::info!(
-                                "{}: Configuration {:?} does not exist, using default",
-                                self.tag(),
-                                T::NAME
-                            );
-                            let mut val = T::default();
-                            val.set_ctx(CfgContext(self.0));
-                            store.borrow_mut().insert(key, Box::leak(Box::new(val)));
-                            Cfg(
-                                UnsafeCell::new(
-                                    store
-                                        .borrow()
-                                        .get(&key)
-                                        .unwrap()
-                                        .downcast_ref()
-                                        .unwrap(),
-                                ),
-                                PhantomData,
-                            )
-                        }
-                    })
-                },
-                |v| Cfg(UnsafeCell::new(v), PhantomData),
-            )
+        get(self.0.as_ref())
     }
 }
 
 impl Default for SharedCfg {
     #[inline]
     fn default() -> Self {
-        Self(DEFAULT_CFG.with(|cfg| *cfg))
+        Self(DEFAULT_CFG.with(Clone::clone))
     }
 }
 
@@ -256,44 +252,79 @@ impl<T: Configuration> From<SharedCfg> for Cfg<T> {
 }
 
 impl SharedCfgBuilder {
-    fn new(tag: String) -> SharedCfgBuilder {
-        let storage = Box::into_raw(Box::new(Storage::new(tag, true)));
-        unsafe {
-            SharedCfgBuilder {
-                ctx: CfgContext(storage.as_ref().unwrap()),
-                storage: Some(Box::from_raw(storage)),
-            }
-        }
+    fn new(tag: &'static str) -> SharedCfgBuilder {
+        let mut storage = Arc::new(Storage::new(tag, true));
+        let ctx = CfgContext(Arc::as_ptr(&storage));
+        Arc::get_mut(&mut storage).unwrap().ctx = CfgContext(ctx.0);
+
+        SharedCfgBuilder { ctx, storage }
     }
 
     #[must_use]
+    #[allow(clippy::missing_panics_doc)]
     /// Insert a type into this configuration.
     ///
     /// If a config of this type already existed, it will
     /// be replaced.
     pub fn add<T: Configuration>(mut self, mut val: T) -> Self {
-        val.set_ctx(self.ctx.clone());
-        if let Some(st) = self.storage.as_mut() {
-            st.data.insert(TypeId::of::<T>(), Box::new(val));
-        }
+        val.set_ctx(CfgContext(self.ctx.0));
+        Arc::get_mut(&mut self.storage)
+            .unwrap()
+            .data
+            .insert(TypeId::of::<T>(), Arc::new(val));
         self
-    }
-}
-
-impl Drop for SharedCfgBuilder {
-    fn drop(&mut self) {
-        if let Some(mut st) = self.storage.take() {
-            st.building = false;
-            let _ = Box::leak(st);
-        }
     }
 }
 
 impl From<SharedCfgBuilder> for SharedCfg {
     fn from(mut cfg: SharedCfgBuilder) -> SharedCfg {
-        let mut st = cfg.storage.take().unwrap();
+        let st = Arc::get_mut(&mut cfg.storage).unwrap();
         st.building = false;
-        SharedCfg(Box::leak(st))
+        SharedCfg(cfg.storage)
+    }
+}
+
+fn get<T>(st: &Storage) -> Cfg<T>
+where
+    T: Configuration,
+{
+    assert!(
+        !st.building,
+        "{}: Cannot access shared config while building",
+        st.tag
+    );
+
+    let tp = TypeId::of::<T>();
+    if let Some(arc) = st.data.get(&tp) {
+        Cfg(
+            UnsafeCell::new(Arc::into_raw(arc.clone().downcast::<T>().unwrap())),
+            PhantomData,
+        )
+    } else {
+        MAPPING.with(|store| {
+            let key = (st.id, tp);
+            if let Some(arc) = store.borrow().get(&key) {
+                Cfg(
+                    UnsafeCell::new(Arc::into_raw(arc.clone().downcast::<T>().unwrap())),
+                    PhantomData,
+                )
+            } else {
+                log::info!(
+                    "{}: Configuration {:?} does not exist, using default",
+                    st.tag,
+                    T::NAME
+                );
+                // store Storage ref, otherwise it could be deallocated
+                let ctx = CfgContext(st.ctx.0);
+                REFS.with(|refs| refs.borrow_mut().insert(ctx.shared()));
+
+                let mut val = T::default();
+                val.set_ctx(ctx);
+                let val = Arc::new(val);
+                store.borrow_mut().insert(key, val.clone());
+                Cfg(UnsafeCell::new(Arc::into_raw(val)), PhantomData)
+            }
+        })
     }
 }
 
@@ -360,12 +391,10 @@ mod tests {
         let t: Cfg<TestCfg> = Cfg::default();
         assert_eq!(t.tag(), "--");
         assert_eq!(t.ctx().id(), t.id());
-        assert_eq!(t.ctx().id(), t.ctx().clone().id());
 
         let t: Cfg<TestCfg> = t.ctx().get();
         assert_eq!(t.tag(), "--");
         assert_eq!(t.ctx().id(), t.id());
-        assert_eq!(t.ctx().id(), t.ctx().clone().id());
 
         let cfg: SharedCfg = SharedCfg::new("TEST2").into();
         let t = cfg.get::<TestCfg>();